Add a customization option for DocBook for a piece, not a book
* tp/Texinfo/Convert/DocBook.pm (_convert),
tp/Texinfo/options_data.txt: add _DOCBOOK_PIECE option.  If set, the
output is not a full book, but a piece of DocBook.  In that case,
before_node_section is not ignored.

* tp/t/test_utils.pl (convert_to_docbook): set _DOCBOOK_PIECE if the
tree has only one element, which should then be before_node_section.
This replaces code that did not follow the API.