* tp/Texinfo/XS/main/build_perl_info.c (build_document): reuse Perl document associated to C document if possible.
* tp/Texinfo/XS/main/build_perl_info.c (build_minimal_document)
(fill_document_hv): bug messages if the document hv is already
set in build_minimal_document.  Bug message if the document hv
is set to another value than the C document associated hv in
fill_document_hv, since it should not happen anymore.

* tp/Texinfo/XS/main/build_perl_info.c: remove rebuild_document.  It
could now be implemented as a wrapper around build_document, and it is
unused.