* doc/texinfo.texi (Adding Output Formats): precise that the Texinfo::Convert::Converter documentation is for backends in Perl.
* doc/texinfo.texi (History): describe the change in 2012 in the past.
Update with information on the use of C.

* doc/texinfo.texi (Generic Translator @command{texi2any}): describe
that two implementations, in Perl and in C exist and how to install
teh C implementation.

* doc/texinfo.texi (@command{texi2any} Environment Variables): update
to clearly explain the differences between the Perl and C
implementations for environment variables.  Document
TEXINFO_C_EMBED_PERL.

* doc/texinfo.texi (Global Customization Variables): explain that
DUMP_TREE is not available in th eC implementation.  Expand on TEST
description.