Macro argument space trimming
* doc/texinfo.tex (\eatleadingcr): replace with implementation
that deals with an empty argument and only handles an active
newline.
(\parsemargdef): do not use an \xempty macro in the definition
of the macro arguments.
* doc/texinfo-tex-test.texi: Update test case.  This means that
spaces are once again trimmed from the end of macro arguments.

\xempty could appear in macro expansions and be undefined.  Report
from Werner.  Rather than providing it with a definition, it is
simpler to get rid of it.