Warn about up pointer for node missing from menu
* tta/C/structuring_transfo/structuring.c (check_node_tree_menu_structure),
* tta/perl/Texinfo/Structuring.pm (check_node_tree_menu_structure)
[CHECK_MISSING_MENU_ENTRY]: If warning issued about a node being missing
from a menu, also check if the up pointer for the node points to a
different node, in order to give more context to the warning.

This is the case for some nodes in the 'libc' manual where some
subsidiary nodes are put at higher sectioning levels than those of their
superior nodes.