Skip unknown tag in Info tag table
* info/nodes.c (get_nodes_of_tags_table): skip an unknown tag in tag
table instead of considering that this is the end of the tag table.
An unknown tag is detected as a line with a colon :.

Note that previous Info readers will not find the nodes and anchors
after an unknown tag, therefore new tags could be introduced
in a backawrd compatible way only when all the older Info readers have
been replaced.  The Emacs Info reader already skips unknown tags.