* tta/C/parsetexi/macro.c (handle_macro), tta/C/parsetexi/parser.c (process_remaining_on_line), tta/perl/Texinfo/ParserNonXS.pm (_handle_macro, _process_remaining_on_line): pass amias information to handle_macro and set the information in this function.
* tta/C/parsetexi/macro.c (handle_macro),
tta/perl/Texinfo/ParserNonXS.pmm (_handle_macro): keep the
macro_call_element in the tree in source mark even if the macro body
is not expanded, in case there are source marks in the
macro_call_element.  If the macrobody is not expanded the state of the
source mark is not set to start, and there is no source mark for an
end of the macro call added.

* tta/perl/t/60macro.t: add/improve comments explaining why
incorrectly placed source marks are correct.