ctexi2any remove use of getcwd
* tta/C/texi2any.c (main): If TEXINFO_DEV_SOURCE is not set, check if
program is uninstalled by checking directory above program directory
in argv[0] contains a file that should be in that directory.  argv[0]
should always contain a directory name because the executable is run
via the libtool wrapper script.
This is simpler, and avoids calling getcwd, which requires gnulib
modules to be installed.