free_node instead of free
* info/session-cmd.c (dump_node_to_stream):
Consistently call 'free_node' instead of 'free' on NODE
object.  It shouuld not make a difference as the node was
returned by info_get_node_of_file_buffer and is a replica
of an object held in the file buffer data structures, but
it should be clearer that the code is correct for somebody
reading it.